Fix StockGenerator_Clothes patch performance
The current StockGenerator_Clothes_HandlesThingDef_Patch implementation
imposes a prohibitive performance overhead when a trader is active on
the map due to it iterating across the entire def database every time
in a high call frequency patch. As a fix, cache the set of defs to be
matched at the patching stage.

This patch was actually a no-op because __result was not declared as a
reference parameter, so the original return value wouldn't have been
modified, so fix that as well.

Also make the build more portable by converting some reference path to
relative ones so that they do not assume a particular install directory
for the game.